add custom code for xpn resources
