--- libraries/google_compute_project.rb
@@ -1,85 +1,0 @@
-# frozen_string_literal: false
-
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-require 'gcp_backend'
-require 'google/compute/property/project_common_instance_metadata'
-require 'google/compute/property/project_common_instance_metadata_items'
-require 'google/compute/property/project_quotas'
-require 'google/compute/property/project_usage_export_location'
-
-# A provider to manage Compute Engine resources.
-class ComputeProject < GcpResourceBase
-  name 'google_compute_project'
-  desc 'Project'
-  supports platform: 'gcp'
-
-  attr_reader :params
-  attr_reader :kind
-  attr_reader :id
-  attr_reader :creation_timestamp
-  attr_reader :name
-  attr_reader :description
-  attr_reader :common_instance_metadata
-  attr_reader :quotas
-  attr_reader :usage_export_location
-  attr_reader :enabled_features
-  attr_reader :self_link
-  attr_reader :default_service_account
-  attr_reader :xpn_project_status
-  attr_reader :default_network_tier
-  attr_reader :vm_dns_setting
-
-  def initialize(params)
-    super(params.merge({ use_http_transport: true }))
-    @params = params
-    @fetched = @connection.fetch(product_url(params[:beta]), resource_base_url, params, 'Get')
-    parse unless @fetched.nil?
-  end
-
-  def parse
-    @kind = @fetched['kind']
-    @id = @fetched['id']
-    @creation_timestamp = @fetched['creationTimestamp']
-    @name = @fetched['name']
-    @description = @fetched['description']
-    @common_instance_metadata = GoogleInSpec::Compute::Property::ProjectCommonInstanceMetadata.new(@fetched['commonInstanceMetadata'], to_s)
-    @quotas = GoogleInSpec::Compute::Property::ProjectQuotasArray.parse(@fetched['quotas'], to_s)
-    @usage_export_location = GoogleInSpec::Compute::Property::ProjectUsageExportLocation.new(@fetched['usageExportLocation'], to_s)
-    @enabled_features = @fetched['enabledFeatures']
-    @self_link = @fetched['selfLink']
-    @default_service_account = @fetched['defaultServiceAccount']
-    @xpn_project_status = @fetched['xpnProjectStatus']
-    @default_network_tier = @fetched['defaultNetworkTier']
-    @vm_dns_setting = @fetched['vmDnsSetting']
-  end
-
-  def exists?
-    !@fetched.nil?
-  end
-
-  def to_s
-    "Project #{@params[:project]}"
-  end
-
-  private
-
-  def product_url(_ = nil)
-    'https://compute.googleapis.com/compute/v1/'
-  end
-
-  def resource_base_url
-    'projects/{{project}}'
-  end
-end
