--- test/integration/verify/controls/google_compute_project.rb
@@ -1,56 +1,0 @@
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-
-title 'Test GCP google_compute_project resource.'
-
-gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
-
-  project = input('project', value: {
-  "network_endpoint_group": "value_networkendpointgroup",
-  "project": "value_project",
-  "region": "value_region",
-  "kind": "value_kind",
-  "id": "value_id",
-  "creation_timestamp": "value_creationtimestamp",
-  "name": "value_name",
-  "description": "value_description",
-  "self_link": "value_selflink",
-  "default_service_account": "value_defaultserviceaccount",
-  "xpn_project_status": "value_xpnprojectstatus",
-  "default_network_tier": "value_defaultnetworktier",
-  "vm_dns_setting": "value_vmdnssetting"
-}, description: 'project description')
-control 'google_compute_project-1.0' do
-  impact 1.0
-  title 'google_compute_project resource test'
-
-  describe google_compute_project(project: gcp_project_id) do
-  	it { should exist }
-  	its('kind') { should cmp project['kind'] }
-  	its('id') { should cmp project['id'] }
-  	its('creation_timestamp') { should cmp project['creation_timestamp'] }
-  	its('name') { should cmp project['name'] }
-  	its('description') { should cmp project['description'] }
-  	its('self_link') { should cmp project['self_link'] }
-  	its('default_service_account') { should cmp project['default_service_account'] }
-  	its('xpn_project_status') { should cmp project['xpn_project_status'] }
-  	its('default_network_tier') { should cmp project['default_network_tier'] }
-  	its('vm_dns_setting') { should cmp project['vm_dns_setting'] }
-
-  end
-
-  describe google_compute_project(project: gcp_project_id) do
-  	it { should_not exist }
-  end
-end
