--- test/integration/verify/controls/google_compute_projects.rb
@@ -1,41 +1,0 @@
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-
-title 'Test GCP google_compute_projects resource.'
-
-gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
-
-  project = input('project', value: {
-  "network_endpoint_group": "value_networkendpointgroup",
-  "project": "value_project",
-  "region": "value_region",
-  "kind": "value_kind",
-  "id": "value_id",
-  "creation_timestamp": "value_creationtimestamp",
-  "name": "value_name",
-  "description": "value_description",
-  "self_link": "value_selflink",
-  "default_service_account": "value_defaultserviceaccount",
-  "xpn_project_status": "value_xpnprojectstatus",
-  "default_network_tier": "value_defaultnetworktier",
-  "vm_dns_setting": "value_vmdnssetting"
-}, description: 'project description')
-control 'google_compute_projects-1.0' do
-  impact 1.0
-  title 'google_compute_projects resource test'
-
-      describe google_compute_projects(project: gcp_project_id, region: project['region']) do
-      it { should exist }
-    end
-end
